modify: sizing API polling

+++ src/main/java/com/tms/backend/dto/SizingStatusDTO.java
@@ -1,3 +1,3 @@
 package com.tms.backend.dto;
 
-public record SizingStatusDTO(String status, int progressPercent, JobAnalysisResponseDTO result) {}
+public record SizingStatusDTO(String status, double progressPercent, JobAnalysisResponseDTO result) {}
