+++ src/main/java/com/tms/backend/tomato/SizingPollStatus.java
@@ -2,6 +2,6 @@ package com.tms.backend.tomato;
 
 import com.tms.backend.dto.TomatoSizingResponse;
 
-public record SizingPollStatus(int progressPercent, TomatoSizingResponse result) {
+public record SizingPollStatus(double progressPercent, TomatoSizingResponse result) {
     public boolean isCompleted() { return result != null; }
 }
